Fix button input detect while game is paused
I fixed the problem where the trick buttons and regular input were still being accepted while the game was paused. This problem also saw the score still being incremented when in the pause menus. The fix was to just add another bool check for the MenusController GameIsPaused bool to see if it was false before allowing score increment and button pressing.